requirements.txt
fastapi
uvicorn
jinja2
psycopg2
python-multipart
scikit-surprise
numpy
pandas